Move FileManagerItemsProvider to Infrastructure
