+++ examples/demo/FluentUI.Demo.Shared/Infrastructure/FileManagerItemsProvider.cs
@@ -2,9 +2,14 @@
 // MIT License - Copyright (c) Microsoft Corporation. All rights reserved.
 // ------------------------------------------------------------------------
 
+
+// ------------------------------------------------------------------------
+// MIT License - Copyright (c) Microsoft Corporation. All rights reserved.
+// ------------------------------------------------------------------------
+
 using FluentUI.Blazor.Community.Components;
 
-namespace FluentUI.Demo.Shared.Providers;
+namespace FluentUI.Demo.Shared.Infrastructure;
 
 public class FileManagerItemsProvider
     : IFileManagerItemsProvider<NoFileEntryData>
